+++ libs/utils/src/index.ts
@@ -1,5 +1,6 @@
 export * from './lib/calcBroadcasterSettlementTotalInfo';
 export * from './lib/calcPgCommission';
 export * from './lib/checkOrderDuringLiveShopping';
+export * from './lib/fileUtil';
 export * from './lib/getHostUrl';
 export * from './lib/kkshowMainDataJsonToObject';
